feat: add dialog, export, and menu action message handlers
