+++ crates/tss-gui/src/handler/mod.rs
@@ -36,8 +36,11 @@
 //! }
 //! ```
 
+pub mod dialog;
 pub mod domain_editor;
+pub mod export;
 pub mod home;
+pub mod menu;
 pub mod source_assignment;
 
 use iced::Task;
@@ -46,8 +49,11 @@ use crate::message::Message;
 use crate::state::AppState;
 
 // Re-export handlers
+pub use dialog::DialogHandler;
 pub use domain_editor::DomainEditorHandler;
+pub use export::ExportHandler;
 pub use home::HomeHandler;
+pub use menu::{MenuActionHandler, MenuMessageHandler};
 pub use source_assignment::SourceAssignmentHandler;
 
 /// Trait for handling messages in the Iced architecture.
